Move core functionality of Config to SimpleSection
